Refactor entidad y agregar capa de servicio
