+++ src/main/java/com/example/catalogo/service/IItemService.java
@@ -1,0 +1,10 @@
+package com.example.catalogo.service;
+
+import com.example.catalogo.entity.Item;
+
+import java.util.List;
+
+public interface IItemService {
+    List<Item> findAll();
+    Item save(Item item);
+}
